02-app/app/requirements.txt
psycopg2==2.9.5
fastapi==0.88.0
python_dotenv==0.21.0